requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Graph & ML
networkx==3.2.1
scikit-learn==1.4.0
hdbscan==0.8.33
sentence-transformers==2.3.1
numpy==1.26.3

# NER
spacy==3.7.2
spacy-lookups-data==1.0.5

# Utils
python-dateutil==2.8.2
requests==2.31.0
google-generativeai==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
